requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiofiles>=23.2.1
google-generativeai>=0.4.0
